small corrections

+++ Database.hpp
@@ -36,9 +36,9 @@ class Database
        
         void loadFromFile();
         void saveToFile();
-        void removeByPersonalID(const unsigned long personalID);
-        void modifySalary(const unsigned long personalID);
-        void modifyAdress(const unsigned long personalID);
-        bool validatePersonalID(const unsigned long personalID);
+        void removeByPersonalID(const unsigned long long personalID);
+        void modifySalary(const unsigned long long personalID);
+        void modifyAdress(const unsigned long long personalID);
+        bool validatePersonalID(const unsigned long long personalID);
 };
